refactor: extract frontend config module
Jira: AI-928
