+++ config/debride.whitelist
@@ -945,14 +945,14 @@ TimeMachineUrlHelper#chapter_on_date_path
 TimeMachineUrlHelper#commodity_on_date_path
 TimeMachineUrlHelper#heading_on_date_path
 TimeMachineUrlHelper#subheading_on_date_path
-TradeTariffFrontend#developer_portal_url
-TradeTariffFrontend#flagsmith_api_url
-TradeTariffFrontend#google_tag_manager_container_id
-TradeTariffFrontend#redis_config
+TradeTariffFrontend::Config#developer_portal_url
+TradeTariffFrontend::Config#flagsmith_api_url
+TradeTariffFrontend::Config#google_tag_manager_container_id
+TradeTariffFrontend::Config#redis_config
 TradeTariffFrontend#roo_wizard?
 TradeTariffFrontend#single_trade_window_linking_enabled?
-TradeTariffFrontend#waf_integration_enabled?
-TradeTariffFrontend#webchat_enabled?
+TradeTariffFrontend::Config#waf_integration_enabled?
+TradeTariffFrontend::Config#webchat_enabled?
 TradeTariffFrontend#welsh?
 TradeTariffFrontend::ServiceChooser#uk_host
 TradeTariffFrontend::ServiceChooser#xi_host
