+++ lib/trade_tariff_frontend.rb
@@ -12,190 +12,11 @@ module TradeTariffFrontend
   autoload :ServiceChooser, 'trade_tariff_frontend/service_chooser'
   autoload :ViewContext,    'trade_tariff_frontend/view_context'
 
-  module_function
+  require_relative 'trade_tariff_frontend/config'
+  extend Config
 
-  def production?
-    environment == 'production'
-  end
-
-  def environment
-    ENV.fetch('ENVIRONMENT', 'production')
-  end
-
-  def id_token_cookie_name
-    cookie_name_for('id_token')
-  end
-
-  def refresh_token_cookie_name
-    cookie_name_for('refresh_token')
-  end
-
-  def cookie_name_for(base_name)
-    case environment
-    when 'production'
-      base_name
-    else
-      "#{environment}_#{base_name}"
-    end.to_sym
-  end
-
-  def redis_config
-    { url: ENV['REDIS_URL'], db: 0, id: nil }
-  end
-
-  # Email of the user who receives all info/error notifications, feedback
-  def from_email
-    ENV['TARIFF_FROM_EMAIL']
-  end
-
-  def to_email
-    ENV['TARIFF_TO_EMAIL']
-  end
-
-  def support_email
-    ENV['TARIFF_SUPPORT_EMAIL'].presence || DEFAULT_SUPPORT_EMAIL
-  end
-
-  def enquiries_email
-    DEFAULT_ENQUIRIES_EMAIL
-  end
-
-  def host
-    ENV.fetch('FRONTEND_HOST', 'http://localhost')
-  end
-
-  def developer_portal_url
-    ENV.fetch('DEVELOPER_PORTAL_URL') { "https://hub.#{base_domain}/" }
-  end
-
-  def flagsmith_api_url
-    ENV['FLAGSMITH_API_URL'].presence || FLAGSMITH_API_URLS[environment]
-  end
-
-  def identity_base_url
-    ENV.fetch('IDENTITY_BASE_URL', 'http://localhost:3005')
-  end
-
-  def identity_cookie_domain
-    @identity_cookie_domain ||= if Rails.env.production?
-                                  ".#{base_domain}"
-                                else
-                                  :all
-                                end
-  end
-
-  def base_domain
-    @base_domain ||= begin
-      domain = ENV['GOVUK_APP_DOMAIN']
-
-      unless /(http(s?):).*/.match(domain)
-        domain = "https://#{domain}"
-      end
-
-      URI.parse(domain).host
-    end
-  end
-
-  def backend_base_domain
-    ENV['BACKEND_BASE_DOMAIN']
-  end
-
-  def revision
+  def self.revision
     @revision ||= `cat REVISION 2>/dev/null || echo 'development'`.strip
-  end
-
-  def self.check_duties_service_url
-    ENV.fetch('CHECK_DUTIES_SERVICE_URL', 'https://www.check-duties-customs-exporting-goods.service.gov.uk')
-  end
-
-  def webchat_enabled?
-    webchat_url.present?
-  end
-
-  def webchat_url
-    configured_url = ENV['WEBCHAT_URL']
-    return if configured_url.blank?
-    return configured_url if configured_url.match?(%r{\Ahttps?://})
-
-    URI.join(WEBCHAT_BASE_URL, configured_url).to_s
-  end
-
-  def legacy_results_to_show
-    ENV.fetch('LEGACY_RESULTS_TO_SHOW', '5').to_i
-  end
-
-  def interactive_search_enabled?
-    !production? && !ServiceChooser.xi?
-  end
-
-  def green_lanes_api_token
-    "Bearer #{ENV['GREEN_LANES_API_TOKEN']}"
-  end
-
-  def google_tag_manager_container_id
-    ENV.fetch('GOOGLE_TAG_MANAGER_CONTAINER_ID', '')
-  end
-
-  def waf_integration_url
-    ENV['WAF_INTEGRATION_URL']
-  end
-
-  def waf_integration_enabled?
-    waf_integration_url.present?
-  end
-
-  def basic_session_authentication?
-    @basic_session_authentication ||= basic_session_password.present?
-  end
-
-  def basic_session_password
-    @basic_session_password ||= ENV['BASIC_PASSWORD']
-  end
-
-  def basic_session_passwords
-    @basic_session_passwords ||= basic_session_password.to_s.split(',').map(&:strip).reject(&:blank?)
-  end
-
-  class FilterBadURLEncoding
-    def initialize(app)
-      @app = app
-    end
-
-    def call(env)
-      @query_string = env['QUERY_STRING'].to_s
-      @path_string = env['PATH_INFO'].to_s
-      begin
-        Rack::Utils.parse_nested_query @query_string
-        return bad_request unless @path_string.ascii_only? && @query_string.ascii_only?
-      rescue Rack::Utils::InvalidParameterError
-        return bad_request
-      end
-
-      @app.call(env)
-    end
-
-    def bad_request
-      @status = 400
-      [
-        @status,
-        { 'Content-Type' => 'application/json' },
-        error_object,
-      ]
-    end
-
-    def error_object
-      [
-        {
-          errors: [
-            {
-              status: @status.to_s,
-              title: 'There was a problem with your query',
-              source: { parameter: 'Invalid query string' },
-            },
-          ],
-        }.to_json,
-      ]
-    end
   end
 
   class FeatureUnavailable < StandardError; end
